requirements.txt
pandas
requests
cli_color
tdqm